refactor order

+++ resources/views/component/order/table.blade.php
@@ -1,5 +1,6 @@
 @php
-    $route =Route::currentRouteName()
+    $route =Route::currentRouteName();
+    $status = App\Models\Orders::$status;
 @endphp
 <div class="post-table">
     <form action="{{url()->current()}}">
@@ -16,10 +17,10 @@
         <thead>
         <tr>
             <th>STT</th>
-            <th>Mã đơn</th>
             <th>Khách hàng</th>
             <th>Tên sản phẩm</th>
             <th>Đơn giá</th>
+            <th>Trạng thái</th>
             <th>Hành động</th>
         </tr>
         </thead>
@@ -27,8 +28,8 @@
         @foreach($objs as $obj)
             <tr>
                 <td>{{$loop->index+1}}</td>
-                <td><a href="{{route('order.show',['order'=>$obj->code])}}">#{{$obj->code}}</a></td>
                 <td>
+                    <div><a href="{{route('order.show',['order'=>$obj->code])}}">#{{$obj->code}}</a></div>
                     {{ $obj->author->name }}
                     {{--                    <div class="post-avatar">--}}
                     {{--                        <label--}}
@@ -38,33 +39,38 @@
                     <div>{{ $obj->author->phone }}</div>
                 </td>
 
-                <td valign="top">
+                <td>
                     {{ $obj->product->name }}
                 </td>
-                <td valign="top">
+                <td>
                     {{ moneyFormat($obj->price) }}
                 </td>
+                <td>{{ App\Models\Orders::$status[$obj->status] }}</td>
                 <td>
                     <div class="d-flex-wrap gap-10">
                         @if(Auth::user()->role===ROLE_ADMIN || Auth::user()->id == $obj->seller_id)
-                            <button class="btn--small btn-ajax danger"
-                                    data-url="{{route('order.destroy', $obj['id'])}}" data-method="delete">Xóa
-                            </button>
-                            @if($obj['status'] != STATUS_ORDERED)
-                                <button class="btn--small btn-ajax warning"
-                                        data-url="{{route('order.updateSimple', $obj['id'])}}"
-                                        data-param='{"status":{{ STATUS_PROCESSING }}}'>Đã đặt hàng
+                            @if(Auth::user()->role===ROLE_ADMIN)
+                                <button class="btn btn--small btn-ajax danger"
+                                        data-url="{{route('order.destroy', $obj['id'])}}" data-method="delete">Xóa
                                 </button>
                             @endif
-                            @if($obj['status'] != STATUS_PROCESSING)
-                                <button class="btn--small btn-ajax"
+                            @if($obj->status == STATUS_COMPLETED)
+                                <button class="btn btn--small btn-ajax"
+                                        data-url="{{route('order.updateSimple', $obj['id'])}}"
+                                        data-param='{"status":{{ STATUS_ORDERED }}}'>
+                                    Reset
+                                </button>
+                            @endif
+
+                            @if($obj->status == STATUS_ORDERED)
+                                <button class="btn btn--small btn-ajax"
                                         data-url="{{route('order.updateSimple', $obj['id'])}}"
                                         data-param='{"status":{{ STATUS_PROCESSING }}}'>
                                     Đang xử lý
                                 </button>
                             @endif
-                            @if($obj['status'] != STATUS_COMPLETED)
-                                <button class="btn--small btn-ajax success"
+                            @if($obj->status != STATUS_COMPLETED)
+                                <button class="btn      btn--small btn-ajax "
                                         data-url="{{route('order.updateSimple', $obj['id'])}}"
                                         data-param='{"status":{{ STATUS_COMPLETED }}}'>
                                     Đã hoàn thành
